Extract GroupCommitter into its own file.
Still trying to break store.rs down into more manageable files.
`GroupCommitter` is nice and isolated and can live in its own file.
